requirements.txt
fastapi~=0.116.1
uvicorn[standard]
scikit-learn~=1.7.2
joblib~=1.5.1
python-dotenv
requests
pymongo
motor

pydantic~=2.11.7
pandas~=2.3.1
openpyxl~=3.1.2
geopy~=2.4.1
openmeteo_requests~=1.6.0
requests-cache~=1.2.1
retry-requests~=2.0.0
matplotlib~=3.10.3
seaborn~=0.13.2
folium~=0.20.0
branca~=0.8.1

# FIXED — Python 3.13 requires NumPy 2.x
numpy<2.0.0

protobuf>=3.20,<5
google-genai

# New dependencies for improved backend
structlog~=24.1.0
redis[hiredis]~=5.0.1
httpx~=0.27.0
python-multipart~=0.0.9

# Phase 1: AI Enhancements
xgboost>=2.0.0
shap>=0.43.0

# Phase 2: Time Series Forecasting
tensorflow>=2.15.0
keras>=3.0.0

# Phase 3: Real-time Data & Alerts
APScheduler>=3.10.4
fastapi-mail>=1.4.1
jinja2>=3.1.2
python-telegram-bot>=20.0
httpx>=0.24.0

# Testing
pytest~=8.0.0
pytest-asyncio~=0.23.0

# Retry logic
tenacity~=8.2.3
